# client/public/swagger.0.0.4.yaml
openapi: 3.0.0
info:
  description: Api documentation
  version: 1.0.1
  title: aim-container API
...
servers:
  - url: http://127.0.0.1:43801/{basePath}
    variables:
      basePath:
        default: api/v1
    description: Local env
security:
  - tokenAuth:
paths:
...
  "/tags/update":
    post:
      operationId: updateTag
      summary: Update tag's name or color.
      tags:
        - Tag
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                color:
                  type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
...
  "/tags/{tag_id}":
    get:
      operationId: getTag
      summary: Get the tag with tag_id
      tags:
        - Tag
      parameters:
        - name: tag_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
...
  "/commits/tags/{commit_id}":
    get:
      operationId: getCommitTags
      summary: Get commit tags
      tags:
        - Commit
      parameters:
          - name: commit_id
            in: path
            required: true
            schema:
              type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/commits/tags/update":
    post:
      operationId: updateCommitTag
      summary: Update commit tag
      tags:
        - Commit
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                commit_hash:
                  type: string
                experiment_name:
                  type: string
                tag_id:
                  type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/commits/info/{experiment}/{commit_id}":
    get:
      operationId: getCommitInfo
      summary: Get commit info
      tags:
        - Commit
      parameters:
        - name: commit_id
          in: path
          required: true
          schema:
            type: string
        - name: experiment
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/commits/search/metric?q={query}":
    get:
      operationId: getCommitsMetricsByQuery
      summary: Get runs metrics
      tags:
        - Commit
      parameters:
          - name: query
            in: path
            required: false
            schema:
              type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/commits/search/dictionary?q={query}":
    get:
      operationId: getCommitsDictionariesByQuery
      summary: Get runs dictionaries
      tags:
        - Commit
      parameters:
          - name: query
            in: path
            required: false
            schema:
              type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
...
  "/commits/tf-summary/params/list":
    post:
      operationId: getTFLogParams
      summary: Get parameters of tf summary log
      tags:
        - Commit
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                path:
                  type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/commits/tf-summary/params/update":
    post:
      operationId: postTFLogParams
      summary: Post parameters of tf summary log
      tags:
        - Commit
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                path:
                  type: string
                params:
                  type: string
                parsed_params:
                  type: object
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
...
  "/projects/{experiment_name}/{commit_id}":
    get:
      operationId: getExperiment
      summary: Get project experiment
      tags:
        - Project
      parameters:
        - name: experiment_name
          in: path
          required: true
          schema:
            type: string
        - name: commit_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExperiment"
  "/projects/insight/{insight_name}":
    get:
      operationId: getProjectInsight
      summary: Get project insight
      tags:
        - Project
      parameters:
        - name: insight_name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectInsight"
  "/projects/{experiment_name}/{commit_id}/{path}":
    get:
      operationId: getExperimentComponent
      summary: Get component content
      tags:
        - Project
      parameters:
        - name: experiment_name
          in: path
          required: true
          schema:
            type: string
        - name: commit_id
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    ProjectInsight:
      type: object
    ProjectExperiment:
      type: object
...